Implemented post build action for Xcode also

+++ test/run_tests.sh
@@ -3,6 +3,8 @@ set -e
 set -x
 
 COMPILE_CCONSTRUCT_COMMAND='clang -x c'
+
+
 pushd 01_hello_world
 rm -rf build
 $COMPILE_CCONSTRUCT_COMMAND config.cc -o cconstruct
@@ -12,6 +14,7 @@ xcodebuild -quiet -workspace build/workspace.xcworkspace -scheme hello_world
 ./build/x64/Debug/hello_world
 popd
 
+
 pushd 02_include_folders
 rm -rf build
 $COMPILE_CCONSTRUCT_COMMAND config.cc -o cconstruct
@@ -20,6 +23,7 @@ xcodebuild -quiet -workspace build/workspace.xcworkspace -scheme include_folders
 ./build/x64/Debug/include_folders
 popd
 
+
 pushd 03_library_dependency
 rm -rf build
 $COMPILE_CCONSTRUCT_COMMAND config.cc -o cconstruct
@@ -27,6 +31,7 @@ $COMPILE_CCONSTRUCT_COMMAND config.cc -o cconstruct
 xcodebuild -quiet -workspace build/xcode/library_dependency.xcworkspace -scheme my_binary
 ./build/xcode/x64/Debug/my_binary
 popd
+
 
 pushd 04_preprocessor
 rm -rf build
@@ -37,6 +42,7 @@ xcodebuild -quiet -workspace build/workspace.xcworkspace -scheme preprocessor -c
 xcodebuild -quiet -workspace build/workspace.xcworkspace -scheme preprocessor -configuration Release
 ./build/x64/Release/preprocessor   
 popd
+
 
 set +e
 pushd 05_compile_flags
@@ -51,8 +57,19 @@ then
 fi
 popd
 
-set -e
-
+pushd 06_post_build_action
+rm -rf build
+$COMPILE_CCONSTRUCT_COMMAND config.cc -o cconstruct
+./cconstruct
+xcodebuild -quiet -workspace build/workspace.xcworkspace -scheme post_build_action
+# building should fail due to the unknown post build command
+if [ $? -eq 0 ]
+then
+  exit 1
+fi
+# but the binary should still have been built and be runnable
+./build/x64/Debug/post_build_action
+popd
 
 pushd ../tools
 rm -rf build
